weakblocks: Update purge logic
This now marks all weak chain tips of height >=0 for removal during
the next step (sets the height to -1) and removes those chain tips
that already have a height of -1.

This causes removal of old weak chain tips to happen in two stages:
First step is marking, and second is removal.

This is important to still allow strong block transmission as
a delta block on top of the last weak block.

+++ src/rpc/weakblock.cpp
@@ -33,10 +33,10 @@ UniValue weakchaintips(const UniValue &params, bool fHelp) {
             "weakchaintips\n"
             "\nGives back the current weak chain tips as pairs of (weak block hash, weak chain height), in chronological order\n");
 
-    std::vector<std::pair<uint256, size_t> > wct_heights = weakChainTips();
+    std::vector<std::pair<uint256, int> > wct_heights = weakChainTips();
     UniValue result(UniValue::VARR);
 
-    for (std::pair<uint256, size_t> p : wct_heights) {
+    for (std::pair<uint256, int> p : wct_heights) {
         UniValue entry(UniValue::VARR);
         entry.push_back(p.first.GetHex());
         entry.push_back(p.second);
